backend/requirements.txt
flask
flask-cors
mysql-connector-python
bcrypt


google-generativeai
python-dotenv
